refactor(stm): group IVC and certificate bench code into dedicated bench modules
Move the benchmark façades and the IVC bench CLI out of the production module namespace into dedicated `bench` submodules (`circuits::halo2_ivc::bench::{helpers,cli}` and `circuits::halo2::bench::helpers`), update the harness imports, and narrow the `IvcProverInput` re-export to the `benchmark-internals` feature.
