+++ mithril-stm/src/proof_system/ivc_halo2_snark/mod.rs
@@ -6,6 +6,6 @@ mod prover_setup;
 pub(crate) mod rolling_state;
 pub(crate) mod verifier_setup;
 
-#[cfg(any(test, feature = "benchmark-internals"))]
+#[cfg(feature = "benchmark-internals")]
 pub(crate) use prover_input::IvcProverInput;
 pub(crate) use prover_setup::IvcSnarkProverSetup;
